requirements.txt
Django==2.0.7
django-environ==0.4.5
django-filter==2.0.0
Markdown==2.6.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
